requirements.txt
appdirs==1.4.3
Django==1.8.7
django-braces==1.10.0
django-media-manager==3.4.2
django-simple-captcha==0.5.5
django-suit==0.2.25
django-suit-ckeditor==0.0.2
django-tinymce==2.3.0
mysqlclient==1.3.7
olefile==0.44
packaging==16.8
Pillow==3.2.0
pkg-resources==0.0.0
psycopg2==2.7.1
pyBarcode==0.7
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
